+++ mopidy/__main__.py
@@ -9,6 +9,7 @@ import sys
 import gobject
 gobject.threads_init()
 
+import pkg_resources
 import pykka.debug
 
 
@@ -36,8 +37,7 @@ from mopidy import exceptions, settings
 from mopidy.audio import Audio
 from mopidy.core import Core
 from mopidy.utils import (
-    deps, importing, log, path, process, settings as settings_utils,
-    versioning)
+    deps, log, path, process, settings as settings_utils, versioning)
 
 
 logger = logging.getLogger('mopidy.main')
@@ -54,10 +54,11 @@ def main():
         log.setup_logging(options.verbosity_level, options.save_debug_log)
         check_old_folders()
         setup_settings(options.interactive)
+        extensions = load_extensions()
         audio = setup_audio()
-        backends = setup_backends(audio)
+        backends = setup_backends(extensions, audio)
         core = setup_core(audio, backends)
-        setup_frontends(core)
+        setup_frontends(extensions, core)
         loop.run()
     except exceptions.SettingsError as ex:
         logger.error(ex.message)
@@ -67,9 +68,9 @@ def main():
         logger.exception(ex)
     finally:
         loop.quit()
-        stop_frontends()
+        stop_frontends(extensions)
         stop_core()
-        stop_backends()
+        stop_backends(extensions)
         stop_audio()
         process.stop_remaining_actors()
 
@@ -138,51 +139,88 @@ def setup_settings(interactive):
         sys.exit(1)
 
 
+def load_extensions():
+    extensions = []
+    for entry_point in pkg_resources.iter_entry_points('mopidy.extension'):
+        logger.debug('Loading extension %s', entry_point.name)
+
+        # TODO Filter out disabled extensions
+
+        try:
+            extension_class = entry_point.load()
+        except pkg_resources.DistributionNotFound as ex:
+            logger.info(
+                'Disabled extension %s: Dependency %s not found',
+                entry_point.name, ex)
+            continue
+
+        extension = extension_class()
+
+        # TODO Validate configuration
+
+        try:
+            extension.validate_environment()
+        except exceptions.ExtensionError as ex:
+            logger.info(
+                'Disabled extension: %s (%s)', extension.name, ex.message)
+            continue
+
+        logger.info(
+            'Loaded extension %s: %s %s',
+            entry_point.name, extension.name, extension.version)
+        extensions.append(extension)
+    return extensions
+
+
 def setup_audio():
+    logger.info('Starting Mopidy audio')
     return Audio.start().proxy()
 
 
 def stop_audio():
+    logger.info('Stopping Mopidy audio')
     process.stop_actors_by_class(Audio)
 
 
-def setup_backends(audio):
+def setup_backends(extensions, audio):
+    logger.info('Starting Mopidy backends')
     backends = []
-    for backend_class_name in settings.BACKENDS:
-        backend_class = importing.get_class(backend_class_name)
-        backend = backend_class.start(audio=audio).proxy()
-        backends.append(backend)
+    for extension in extensions:
+        for backend_class in extension.get_backend_classes():
+            backend = backend_class.start(audio=audio).proxy()
+            backends.append(backend)
     return backends
 
 
-def stop_backends():
-    for backend_class_name in settings.BACKENDS:
-        process.stop_actors_by_class(importing.get_class(backend_class_name))
+def stop_backends(extensions):
+    logger.info('Stopping Mopidy backends')
+    for extension in extensions:
+        for backend_class in extension.get_backend_classes():
+            process.stop_actors_by_class(backend_class)
 
 
 def setup_core(audio, backends):
+    logger.info('Starting Mopidy core')
     return Core.start(audio=audio, backends=backends).proxy()
 
 
 def stop_core():
+    logger.info('Stopping Mopidy core')
     process.stop_actors_by_class(Core)
 
 
-def setup_frontends(core):
-    for frontend_class_name in settings.FRONTENDS:
-        try:
-            importing.get_class(frontend_class_name).start(core=core)
-        except exceptions.OptionalDependencyError as ex:
-            logger.info('Disabled: %s (%s)', frontend_class_name, ex)
+def setup_frontends(extensions, core):
+    logger.info('Starting Mopidy frontends')
+    for extension in extensions:
+        for frontend_class in extension.get_frontend_classes():
+            frontend_class.start(core=core)
 
 
-def stop_frontends():
-    for frontend_class_name in settings.FRONTENDS:
-        try:
-            frontend_class = importing.get_class(frontend_class_name)
+def stop_frontends(extensions):
+    logger.info('Stopping Mopidy frontends')
+    for extension in extensions:
+        for frontend_class in extension.get_frontend_classes():
             process.stop_actors_by_class(frontend_class)
-        except exceptions.OptionalDependencyError:
-            pass
 
 
 if __name__ == '__main__':