+++ tests/backends/local/events_test.py
@@ -1,12 +1,12 @@
 from mopidy import settings
-from mopidy.backends.local import LocalBackend
+from mopidy.backends.local import actor
 
 from tests import unittest, path_to_data_dir
 from tests.backends.base import events
 
 
 class LocalBackendEventsTest(events.BackendEventsTest, unittest.TestCase):
-    backend_class = LocalBackend
+    backend_class = actor.LocalBackend
 
     def setUp(self):
         settings.LOCAL_TAG_CACHE_FILE = path_to_data_dir('empty_tag_cache')