+++ tests/backends/local/playback_test.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals
 
 from mopidy import settings
-from mopidy.backends.local import LocalBackend
+from mopidy.backends.local import actor
 from mopidy.core import PlaybackState
 from mopidy.models import Track
 from mopidy.utils.path import path_to_uri
@@ -12,12 +12,11 @@ from tests.backends.local import generate_song
 
 
 class LocalPlaybackControllerTest(PlaybackControllerTest, unittest.TestCase):
-    backend_class = LocalBackend
+    backend_class = actor.LocalBackend
     tracks = [
         Track(uri=generate_song(i), length=4464) for i in range(1, 4)]
 
     def setUp(self):
-        settings.BACKENDS = ('mopidy.backends.local.LocalBackend',)
         settings.LOCAL_TAG_CACHE_FILE = path_to_data_dir('empty_tag_cache')
         super(LocalPlaybackControllerTest, self).setUp()